Future Proofing locations.py
Earlier I thought that preparing for expanding to 100 max rooms was enough, but, just in case, I made it so I can expand to 1000 max rooms. I doubt I'll ever go that far, but I could see ~100 being potentially viable for a long async. Still planning on releasing 1.0.0 with a max of 20.